requirements.txt
# Core dependencies for Halloween Roaster
openai>=1.0.0
picamera2>=0.3.18
pillow>=10.0.0
SpeechRecognition>=3.10.0
pygame>=2.5.0
gTTS>=2.5.0
PyAudio>=0.2.14
opencv-python>=4.8.0
numpy>=1.24.0

# Google Drive integration for trace file uploads
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
